# src/student-letter-submission/dto/update-student-letter-submission.dto.ts
import { PartialType } from '@nestjs/swagger';
import { CreateStudentLetterSubmissionDto } from './create-student-letter-submission.dto';

export class UpdateStudentLetterSubmissionDto extends PartialType(
  CreateStudentLetterSubmissionDto,
) {}