# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/student/{id}": {
      "get": {
        "operationId": "StudentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Student"
        ]
      },
      "patch": {
        "operationId": "StudentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Student"
        ]
      },
      "delete": {
        "operationId": "StudentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Student"
        ]
      }
    },
...
    "/api/letter/{id}": {
      "get": {
        "operationId": "LetterController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Letter"
        ]
      },
      "patch": {
        "operationId": "LetterController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLetterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Letter"
        ]
      },
      "delete": {
        "operationId": "LetterController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Letter"
        ]
      }
    },
...
    "/api/letter-template": {
      "post": {
        "operationId": "LetterTemplateController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLetterTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "LetterTemplate"
        ]
      },
      "get": {
        "operationId": "LetterTemplateController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Nomor halaman yang ingin ditampilkan",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Jumlah data per halaman",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Kata kunci untuk pencarian berdasarkan nama atau deskripsi",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "LetterTemplate"
        ]
      }
    },
    "/api/letter-template/upsert": {
      "post": {
        "operationId": "LetterTemplateController_upsert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLetterTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "LetterTemplate"
        ]
      }
    },
    "/api/letter-template/{id}": {
      "get": {
        "operationId": "LetterTemplateController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "LetterTemplate"
        ]
      },
      "patch": {
        "operationId": "LetterTemplateController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLetterTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "LetterTemplate"
        ]
      },
      "delete": {
        "operationId": "LetterTemplateController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "LetterTemplate"
        ]
      }
    },
...
    "/api/student-letter-submission": {
      "post": {
        "operationId": "StudentLetterSubmissionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentLetterSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "StudentLetterSubmission"
        ]
      },
      "get": {
        "operationId": "StudentLetterSubmissionController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Nomor halaman yang ingin ditampilkan",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Jumlah data per halaman",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Kata kunci untuk pencarian berdasarkan nama atau deskripsi",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "StudentLetterSubmission"
        ]
      }
    },
    "/api/student-letter-submission/{id}/print": {
      "get": {
        "operationId": "StudentLetterSubmissionController_print",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "StudentLetterSubmission"
        ]
      }
    },
    "/api/student-letter-submission/{id}": {
      "get": {
        "operationId": "StudentLetterSubmissionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "StudentLetterSubmission"
        ]
      },
      "patch": {
        "operationId": "StudentLetterSubmissionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentLetterSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "StudentLetterSubmission"
        ]
      },
      "delete": {
        "operationId": "StudentLetterSubmissionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "StudentLetterSubmission"
        ]
      }
    },
...
  },
  "info": {
    "title": "DIGILET",
    "description": "Dokumentasi API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "UpdateStudentDto": {
        "type": "object",
        "properties": {
          "fullname": {
            "type": "string",
            "example": "John Doe",
            "description": "Full name"
          },
          "studyProgramId": {
            "type": "number",
            "example": 1,
            "description": "Study program ID"
          },
          "email": {
            "type": "string",
            "example": "admin@example.com",
            "description": "email"
          },
          "address": {
            "type": "string",
            "example": "Jl. Merdeka No. 123",
            "description": "Address"
          },
          "nim": {
            "type": "string",
            "example": "1234567890",
            "description": "NIM"
          },
          "classYear": {
            "type": "string",
            "example": "2021",
            "description": "Class year"
          },
          "phoneNumber": {
            "type": "string",
            "example": "081234567890",
            "description": "Handphone number"
          },
          "birthday": {
            "type": "string",
            "example": "2000-01-01",
            "description": "Birthday"
          },
          "birthplace": {
            "type": "string",
            "example": "Jakarta",
            "description": "Birthplace"
          },
          "gender": {
            "type": "string",
            "example": "Laki-laki/Perempuan",
            "description": "Gender"
          }
        }
      },
...
      "UpdateLetterDto": {
        "type": "object",
        "properties": {
          "letterName": {
            "type": "string",
            "example": "Surat Pengantar",
            "description": "nama surat"
          },
          "referenceNumber": {
            "type": "string",
            "example": "123/456/789",
            "description": "nomor surat"
          },
          "expiredDate": {
            "type": "number",
            "example": "2023-09-10",
            "description": "tanggal kedaluwarsa"
          },
          "letterNumberingStart": {
            "type": "number",
            "example": 1,
            "description": "nomor urut surat"
          },
          "category": {
            "type": "string",
            "description": "kategori surat",
            "enum": [
              "faculty",
              "study_program",
              "university",
              "all"
            ]
          },
          "signatureType": {
            "type": "string",
            "description": "jenis tanda tangan",
            "enum": [
              "barcode",
              "digital"
            ]
          }
        }
      },
...
      "CreateLetterTemplateDto": {
        "type": "object",
        "properties": {
          "letterId": {
            "type": "number",
            "example": 1,
            "description": "ID surat"
          },
          "content": {
            "type": "string",
            "example": "content",
            "description": "Content"
          }
        },
        "required": [
          "letterId",
          "content"
        ]
      },
      "UpdateLetterTemplateDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateStudentLetterSubmissionDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "number",
            "description": "ID dari student"
          },
          "letterId": {
            "type": "number",
            "description": "ID dari letter"
          },
          "letterNumber": {
            "type": "string",
            "description": "Nomor surat"
          },
          "status": {
            "type": "string",
            "description": "Status surat",
            "enum": [
              "pending",
              "waiting_signature",
              "approved",
              "rejected"
            ],
            "default": "pending"
          }
        },
        "required": [
          "studentId",
          "letterId",
          "status"
        ]
      },
      "UpdateStudentLetterSubmissionDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "number",
            "description": "ID dari student"
          },
          "letterId": {
            "type": "number",
            "description": "ID dari letter"
          },
          "letterNumber": {
            "type": "string",
            "description": "Nomor surat"
          },
          "status": {
            "type": "string",
            "description": "Status surat",
            "enum": [
              "pending",
              "waiting_signature",
              "approved",
              "rejected"
            ],
            "default": "pending"
          }
        }
      },
...
    }
  }
}